refactor: Renamed model file names
